+++ src/app/modules/error/not-found-msg-switch.component.scss
@@ -1,0 +1,3 @@
+:host {
+    display: block;
+}
