+++ src/app/modules/error/not-found.component.pug
@@ -1,14 +1,7 @@
 .header
     mat-icon help_outline
 h3 Not Found
-div.info([ngSwitch]="errorTypeObservable|async")
-    ng-container(*ngSwitchCase="1")
-        p The passage could not be found. It either expired or has yet to start.
-        p Please select another passage.
-    ng-container(*ngSwitchCase="2")
-        p The requested vehicle could not be found. It might not be active at the moment.
-    ng-container(*ngSwitchDefault)
-        p The requested resource can not be found.
+app-not-found-msg-switch
 mat-nav-list(role="list")
     mat-list-item(*ngFor="let endpoint of endpoints",role="listitem"[routerLink]="[endpoint.path]")
         mat-icon(matListIcon) {{endpoint.icon}}
